Move cmd/ndt7-client/internal packages into top level internal.
Make ndt7-prometheus-exporter directory a sibling of ndt7-client.
