+++ internal/emitter/json.go
