+++ internal/emitter/quiet.go
