+++ internal/emitter/summary.go
