+++ internal/emitter/summary_test.go
